plant-diagnosis-model/requirements.txt
torch
torchvision
timm
pyyaml
scikit-learn
numpy
Pillow
rich

